revise the gassfh, correct the SF length for current snapshot As described in Vivienne's email: """ In line 255 and 205 where we identify all the SSPs that fall in the time step of interest (ind_all)
For the current snapshot i.e. j=i  we only identify SSPs that are younger than 10^7 years (i.e. delta_age/2.)
But we include them into the gassfh and newstarsfh array weighted by delta_age (lines 210 and 261). I think this will double the weight given to the SFR in the current snapshot compared to what we want….
"""